Added knight and part of king moves.
