+++ src/fens.rs
@@ -1,5 +1,5 @@
 pub const FEN_START: &str = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";
-pub const FEN_TRICKY: &str = "r3k2r/p1ppqpb1/bn2pnp1/3PN3/1p2P3/2N2Q1p/PPPBBPPP/R3K2R w KQkq - 0 1";
+pub const FEN_TRICKY: &str = "r3k2r/p1ppqpb1/bn2pnp1/3PN3/1p2P3/2N2Q1p/PPPBBPPP/R3K2R b KQkq - 0 1";
 pub const FEN_ENPASSANT: &str = "rnbqkbnr/p1p1pppp/8/1pP4P/3p4/4P3/PP1P1PP1/RNBQKBNR w KQkq b6 0 1";
 pub const FEN_PROMOTION_WHITE: &str = "4r3/5P2/8/1k6/8/8/5p2/1K4N1 w - - 0 1";
 pub const FEN_PROMOTION_BLACK: &str = "4r3/5P2/8/1k6/8/8/5p2/1K4N1 b - - 0 1";
